[ATOMIC][CHG] change CAS, not need old ptr (just value)

+++ src/fifo/test/main.c
@@ -17,7 +17,7 @@ struct job {
 void *worker(void *arg)
 {
     struct job *job = arg;
-    int i = pthread_self() % 23, j = -1;
+    int i = (int)pthread_self() % 23, j = -1;
     int tries = 10;
     int total_pop = 0, total_push = 0;
 
@@ -26,23 +26,23 @@ void *worker(void *arg)
 
     while (tries-- && cfifo_size(job->results) < job->min_results) {
         if (cfifo_push(job->results, (void *)(intptr_t)(++i)) != 0) {
-            printf("(%d) push error\n", pthread_self());
+            printf("(%d) push error\n", (int)pthread_self());
             abort();
         } else {
-            // printf("(%d) push (%d)\n", pthread_self(), i);
+            // printf("(%d) push (%d)\n", (int)pthread_self(), i);
             total_push += i;
         }
 
         if ((j = (int)(intptr_t)cfifo_pop(job->results)) == (intptr_t)NULL) {
-            // printf("(%d) pop error\n", pthread_self());
+            // printf("(%d) pop error\n", (int)pthread_self());
             continue;
         } else {
-            // printf("(%d) pop (%d)\n", pthread_self(), j);
+            // printf("(%d) pop (%d)\n", (int)pthread_self(), j);
             total_pop += j;
         }
     }
 
-    printf("(%d) total: push (%d), pop (%d)\n", pthread_self(),
+    printf("(%d) total: push (%d), pop (%d)\n", (int)pthread_self(),
            total_push, total_pop);
 
     return (void *)0;
